[Doc][Example][Bugfix] Elements in local_device_ids should be casted … (#3782)
### What this PR does / why we need it?
It's a tiny bugfix in the `gen_ranktable.py` script. The script is an
util to help setup an example case. It is used to prepare a ranktable
before disaggregated prefill deployment.

Elements in `local_device_ids` list should be casted to `int` type
before referred for a MOD math operation.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
No.

- vLLM version: v0.11.0
- vLLM main:
https://github.com/vllm-project/vllm/commit/c9461e05a4ed3557cfbf4b15ded1e26761cc39ca

---------

Signed-off-by: paulyu12 <507435917@qq.com>
Signed-off-by: xqchen7 <chenxueqing7@huawei.com>